correct camera_info reading

+++ src/ar_core/AugmentedCamera.h
@@ -66,8 +66,8 @@ private:
     KDL::Frame                  world_to_cam_tr;
     bool                        is_pose_from_subscriber =true;
 
-    cv::Mat                     camera_matrix= cv::Mat::zeros(3, 3, CV_32F);
-    cv::Mat                     camera_distortion= cv::Mat::zeros(1, 5, CV_32F);
+    cv::Mat                     camera_matrix= cv::Mat::zeros(3, 3, CV_64F);
+    cv::Mat                     camera_distortion= cv::Mat::zeros(1, 5, CV_64F);
 
     // pose estimation
     cv::Ptr<cv::aruco::CharucoBoard>   charuco_board;
